requirements.txt
boto3==1.34.0
openai>=1.12.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
langgraph==0.2.0
langchain==0.2.0
psycopg2-binary==2.9.9
chromadb==0.4.22
sqlalchemy==2.0.25
alembic==1.13.1
stripe==7.8.0
requests==2.31.0
# Performance optimizations
aiohttp==3.9.1
aiodns==3.1.1
redis==5.0.1
orjson>=3.9.14 
# Document processing
python-docx==1.1.2 